Store stream windows as deltas from the initial value (and use i64).
Kinda weird that the spec requires updates to initial_window_size to affect
existing streams as well tbh.

Could probably make do with 32-bit integers, but that requires some overflow
checking, and I'm too lazy for that.

+++ cno/core.h
@@ -161,8 +161,8 @@ struct cno_stream_t
     uint32_t id;
     uint8_t /* enum CNO_STREAM_ACCEPT */ accept;
     uint8_t /* enum CNO_STREAM_FLAGS  */ flags;
-     int32_t window_recv;
-     int32_t window_send;
+     int64_t window_recv;
+     int64_t window_send;
     uint64_t remaining_payload;
 };
 
@@ -194,8 +194,8 @@ struct cno_connection_t
     uint8_t  continued_flags;
     uint32_t continued_stream;
     uint32_t continued_promise;
-     int32_t window_recv;
-     int32_t window_send;
+     int64_t window_recv;
+     int64_t window_send;
     uint32_t last_stream[2]; // dereferencable with CNO_REMOTE/CNO_LOCAL
     uint32_t stream_count[2];
     uint32_t goaway_sent;
